Model Constructors

StateSpaceSciML provides a set of functions to construct NODE and UDE with varying levels of customization. The model constructors all require the data to be passed using a DataFrame object from the DataFrames.jl library. The data frame should be organized with a column for time named t and the remianing columns shoud have the value of the state variables at each point in time.

t	$X_1$	$X_2$

0.1	0.0	-1.1

0.2	0.01	-0.9

0.5	0.51	-1.05

Table: Example data set with two state vaiables

Currently mising data are not supported, but irregular intervals between time points are allowed.

Each constructor function will require addtional inputs to specify the model structure. For example, the CustomDerivatives function requires the user supply the known functional forms through the derivs! argument. These aregumetns are described in detail in the subsection for each model type.

Finally, the constructor functions share a set of key work arguments that are used to tune the model fitting procedure. These control the weight given to the process model, observaiton model and regualrization in the loss function and can be tuned to control the complexity of the estimated model and to accomidate variying levels of observaitonal errors:

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=0.0 : Weight given to regularization in the loss function.

In addtion to these weighting parameters two key work arguments l = 0.25 and extrap_rho = 0.0 control how the model extapolated beyond the observed data. The paramter l defines how far away the modle will extrapolate before shifting to the default behavior and extrap_rho defined the default when extrapolating. When forecasting the model will modify the trained process model $f(u_t;\theta)$ when extapolating to a new fucntion

\[bar{f}(u_t|\theta,l,\rho )= \left\{ \begin{array}{ll} f(u_t;\theta) & min(\hat{u}) < u_t < max(\hat{u}) \\ e^{(\frac{u_t - min(\hat{u}_t)}{l})^2}f(u_t;\theta) + (1-e^{(\frac{u_t - min(\hat{u}_t)}{l})^2}) \rho &u_t < min(\hat{u}) \\ e^{(\frac{u_t - max(\hat{u}_t)}{l})^2}f(u_t;\theta) - (1-e^{(\frac{u_t - min(\hat{u}_t)}{l})^2}) \rho &u_t > max(\hat{u}) \\ \end{array} \right. \]

NODES and NNDE

The simplest models to implement are the fully nonparametrics neural ordinary differential equation (NODE) and neural newtork difference equation (NNDE). These functions use a neural network to represent the right hand side of a system of differential equations and differnce equation respectively

\[ \frac{dx}{dt} = NN(x;w,b) \\ x_{t+1} = NN(x_t;w,b)\]

The NODE function builds a NODE model for a user supplied data set and NNDE build

StateSpaceSciML.NODE — Method

NODE(data;kwargs ... )

Constructs a nonparametric continuous time model for the data set data using a single layer neural network to represent the systems dynamics.

Model equations

\[ dx/dt = NN(x;w,b)\]

...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

Covariates can be added to the model by supplying a second data frame X with thier values at differnt points in time. The value of the covaritates are appended to the state vector $x$ to create the input for the neural network

\[ \frac{dx}{dt} = NN(x,X(t);w,b) \\ x_{t+1} = NN(x_t,X_t;w,b)\]

The NODE models use a linear interpolation to approximate the value fo the covarites between observations. odel with covariates have not been developed for the discrete time case yet.

StateSpaceSciML.NODE — Method

NODE(data,X;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argumetn X should have a column for time t with the vlaue fo time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included int he data frame.

When X is provided the derivs function must have the form derivs!(du,u,x,p,t) where x is a vector with the value of the coarates at time t.

UDEs

The CustomDerivatives and CustomDifference function can be used to build models that combine nerual networks and known functional forms. These function take user defined models and consturct a loss function and provide access to the model fitting and testing functions provided by StateSpaceSciML.jl

StateSpaceSciML.CustomDerivatives — Method

CustomDerivatives(data,derivs!,initial_parameters;kwargs ... )

Constructs a UDE model for the data set data based on user defined derivitivs derivs. An initial guess of model parameters are supplied with the initia_parameters argument.

...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

derivs: a Function of the form derivs!(du,u,p,t) where u is the value of the state variables, p are the model parameters, t is time, and du is updated with the value of the derivitives

init_parameters: A NamedTuple with the model parameters. Neural network parameters must be listed under the key NN.

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

StateSpaceSciML.CustomDiffernce — Method

CustomDiffernce(data,step,initial_parameters;kwrags...)

Constructs a UDE model for the data set data based on user defined difference equation step. An initial guess of model parameters are supplied with the initia_parameters argument.

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

step: a Function of the form step(u,t,p) where u is the value of the state variables, p are the model parameters.

init_parameters: A NamedTuple with the model parameters. Neural network parameters must be listed under the key NN.

...

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

Adding Covariates

In this context the

StateSpaceSciML.CustomDerivatives — Method

CustomDerivatives(data,X,derivs!,initial_parameters;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argumetn X should have a column for time t with the vlaue fo time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included int he data frame.

When X is provided the derivs function must have the form derivs!(du,u,x,p,t) where x is a vector with the value of the coarates at time t.

StateSpaceSciML.CustomDiffernce — Method

CustomDiffernce(data,X,step,initial_parameters;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argumetn X should have a column for time t with the vlaue fo time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included int he data frame.

When X is provided the step function must have the form step(u,x,t,p) where x is a vector with the value of the coarates at time t.

Other functions

StateSpaceSciML.NNDE — Method

NNDE(data;kwargs ...)

Constructs a nonparametric discrete time model for the data set data using a single layer neural network to reporesent the systems dynamics.

Model equations

\[x_{t+1} = NN(x_t;w,b)\]

...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

StateSpaceSciML.DiscreteUDE — Method

DiscreteUDE(data,step,init_parameters;kwargs ...)

Constructs an additive UDE model with user supplied difference equations step and a single layer neural network. When init_parameters are provided for the use supplied function their values will be estiated in the training process.

Model equaitons

\[x_{t+1} = f(x_t; heta) + NN(x_t;w,b)\]

...

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

StateSpaceSciML.UDE — Method

UDE(data,derivs,init_parameters;kwargs...)

Constructs an additive continuous time UDE model with user supplied derivitives step and a single layer neural network. When init_parameters are provided for the user supplied function their values will be estiated during model training.

Model equaitons

\[dx/dt = = f(x; heta) + NN(x;w,b)\]

...

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...